+++ src/rogallo/__main__.py
@@ -139,6 +139,7 @@ def show_dignoastics() -> None:
     from cryptography import __version__ as cryptography_version
     from gemtext import __version__ as gemtext_version
     from gophermap import __version__ as gophermap_version
+    from html2gemtext import __version__ as html2gemtext_version
     from port70 import __version__ as port70_version
     from port79 import __version__ as port79_version
     from sybaritic import __version__ as sybaritic_version
@@ -152,6 +153,7 @@ def show_dignoastics() -> None:
     print(f"cryptography: {cryptography_version}")
     print(f"gemtext: {gemtext_version}")
     print(f"gophermap: {gophermap_version}")
+    print(f"html2gemtext: {html2gemtext_version}")
     print(f"port70: {port70_version}")
     print(f"port79: {port79_version}")
     print(f"rogallo: {__version__}")